requirements.txt
async-timeout==4.0.2
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dateutil==2.8.2
python-dotenv==0.21.1
redis==4.5.1
six==1.16.0
Werkzeug==2.2.3
